Add new and refactor existing examples and tests
