+++ examples/_shared/src/chat-room/index.ts
@@ -1,0 +1,3 @@
+export * from './message-list';
+export * from './message-submit';
+export * from './messaging';
