+++ examples/base-path/app/(simple)/api/ws/route.ts
