+++ examples/base-path/next-env.d.ts
