+++ examples/base-path/next.config.mjs
@@ -1,0 +1,3 @@
+export default {
+  basePath: '/some-base-path',
+};
