+++ examples/base-path/tsconfig.json
