--- examples/chat-room-with-custom-server/next.config.js
@@ -1,1 +1,0 @@
-module.exports = {};
