--- examples/chat-room-with-custom-server/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "chat-room-with-custom-server",
-  "private": true,
-  "scripts": {
-    "build": "next build",
-    "start": "NODE_ENV=production tsx server.ts",
-    "dev": "tsx server.ts",
-    "prepare": "next-ws patch"
-  },
-  "dependencies": {
-    "next": "^15.2.3",
-    "next-ws": "workspace:^",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
-  },
-  "devDependencies": {
-    "@types/react": "^19.0.8",
-    "tsx": "^4.19.3"
-  }
-}
