--- examples/chat-room-with-custom-server/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "include": ["app/**/*", "server.ts", ".next/types/**/*.ts"],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "target": "ES2017",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "strict": false,
-    "noEmit": true,
-    "incremental": true,
-    "module": "esnext",
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "plugins": [{ "name": "next" }],
-    "forceConsistentCasingInFileNames": true
-  }
-}
