--- examples/chat-room/app/chat/_shared/message.ts
@@ -1,4 +1,0 @@
-export interface Message {
-  author: string;
-  content: string;
-}
