--- examples/chat-room/next.config.js
@@ -1,1 +1,0 @@
-module.exports = {};
