+++ examples/custom-server/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@examples/chat-room",
+  "name": "@examples/custom-server",
   "private": true,
   "scripts": {
     "build": "next build",
-    "start": "next start",
-    "dev": "next dev",
+    "start": "NODE_ENV=production tsx server.ts",
+    "dev": "tsx server.ts",
     "prepare": "next-ws patch"
   },
   "dependencies": {
@@ -15,6 +15,7 @@
     "shared": "workspace:^"
   },
   "devDependencies": {
-    "@types/react": "^19.0.8"
+    "@types/react": "^19.0.8",
+    "tsx": "^4.19.3"
   }
 }
