+++ playwright.config.ts
@@ -15,10 +15,16 @@ export default defineConfig({
       reuseExistingServer: !process.env.CI,
     },
     {
-      cwd: 'examples/chat-room-with-custom-server',
-      env: { PORT: '3002' },
-      command: 'pnpm dev',
+      cwd: 'examples/base-path',
+      command: 'pnpm dev --port 3002',
       port: 3002,
+      reuseExistingServer: !process.env.CI,
+    },
+    {
+      cwd: 'examples/custom-server',
+      command: 'pnpm dev',
+      env: { PORT: '3003' },
+      port: 3003,
       reuseExistingServer: !process.env.CI,
     },
   ],
